Modified references to lib files.
Because of the new structure, the references to "lib/..." were wrong.
Changed them to "../lib/...".
I looked for a more resilient way to do so, by an absoulte naming from
the root project folder, (like "~/lib/..." or "//lib/...") but it seems
that I only found feature requests about this. It is probably not
implemented yet. :-1:
The file ".project" is for Eclipse.