AStar working as hell. :rocket: :guitar:
I made a lot of changes on the AStar so that it is possible to create
the states along the exploration, but I had to deal with a way to
compare two different instances of a state which would represent the
same state : I used a string representation (hash). It also optimised
because the closed nodes could be erased.
The algorithm is finally simpler and more efficient! ;)
I defined the heuristic for the ontop relation.

Try the example in complex mode!! It will put the yellow brick on the
red plank. Even the professor's version freezes on this one. ;)