The planner is done!
I added the other relations to the heuristic calculation (above, under,
right, left, beside and EVEN holding).
The "holding" constraint is now integrated and working.
I made several tests with several constraints at a time and it works
pretty good! :+1:
The planner is therefore done. One possibility would be to count cost as
number of "actions" instead as number of "moves". The optimisation would
be even better. It is possible to adapt this quite easily. Maybe next
time with Juho, and so we could compare.