Add functionality of simple object descriptions
The interpreter can now handle the verb part, as well as the object part if it'
s simple. So "the red ball" works, but "the ball in the box" is not implemented yet. One form, or one form and optionally colour and/or size, is fine. "the floor" also works. Currently treating "any" and "the" the same, and rejecting everything else.