Make the heuristic only take one node
The heuristic as defined in Astar should take two nodes and give an
approximation of the distance between them. However, we don't have a goal node;
 we have abstract conditions for when a node fulfils the goal. Thus, the
heuristic has to use those abstract conditions directly, instead of comparing
the current node to a goal node which we don't even have explicitly defined.