Changed stuff in the use of the priority queues.
Use a Heap for the closedset with "hash" string (no more Nodes to be
kept in memory moreover), and a priority queue for the openset based on
f_score value. We don't check anymore if a state is in the openset when
adding one but when poping a new state from the openset, we compute it
only if it is not in the closedset.
Compared both implementation on "put all boxes on the floor" : it much
faster!! :+1: